+++ ft_lstlen.c
@@ -6,15 +6,15 @@
 /*   By: oevtushe <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/02 14:19:14 by oevtushe          #+#    #+#             */
-/*   Updated: 2017/11/12 16:28:05 by oevtushe         ###   ########.fr       */
+/*   Updated: 2017/11/16 16:08:10 by oevtushe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_lstlen(const t_list *lst)
+size_t	ft_lstlen(const t_list *lst)
 {
-	int len;
+	size_t	len;
 
 	len = 0;
 	while (lst)
